+++ src/app/(home)/resources/page.tsx
